drop reentry_scan hook
In pip 19, the isolation of the build environment means that
pkg_resources in install hooks provides access only to the distributions
in the build environment (setuptools, wheel, reentry + dependencies).

The `reentry_scan` install hook was therefore no longer able to scan for
entry points in the target environment and lost its purpose.

 * update changelog
 * fix #3: clarify meaning of "extras" in entry points in README