config: add option to set data_filename in config.
Sometimes it can be useful to choose the name of the data
file, instead of letting reentry pick the name by its own.
The new field is optional, which means the behavior will
remain the same, unless the user asks for a diffenrent
data filename in the configuration file or via environment variable.

Signed-off-by: Asaf Kahlon <asafka7@gmail.com>