Fixing collections.abc import
I now fix the collections.abc.Sequence import, that was
being imported as collections.Sequence instead, raising warnings
(and will not work in py 3.10).

I also add a "url" field in the setup.json, so on pypi there will
be a link back to the repository.

Version upped to 1.3.2 to make this ready for a new release.