Do not resolve Python Path symlinks for datafile hash
In some cases, the Python binaries of multiple virtual environments can
be symlinks that point to the same file. In this case, the hash of the
reentry data file is the same for these environments, since
`Path.resolve()` also resolves symlinks.

As `sys.executable` already returns the absolute path to the environment
Python binary, we can just remove `.resolve()` to avoid resolving the
symlinks.